Moved tree into a new module layout
